added filtering logic to file names
this just means i learned regexgit add .! been doing this for almost two years and just now i'm getting around to it, huh. In other news it took forever to get this done because of school and a nagging dad 'get a job noah' welp dad you may just be right.